# specs/002-builder-modelli/contracts/builder-modelli-api.openapi.yaml
openapi: 3.0.3
info:
  title: GEMODO - Builder Modelli
  version: 0.6.0
  description: >
    Creazione e ciclo di vita di modello/versione (007 tasks.md T006). Sostituisce
    integralmente la 0.2.0: quella descriveva tipi documento/categorie come CRUD
    locale del builder, un design pre-010 mai implementato cosi'. Oggi tipi
    documento/categorie sono letti in sola lettura dalla discovery live
    dell'integrazione connessa (vedi integrazioni-api.openapi.yaml, gia'
    pubblicato) - questo contratto copre solo le route realmente implementate in
    backend/app/builder/api.py che quel contratto non gia' copre: lettura
    struttura disponibile per un tipo documento gia' associato, creazione
    modello/versione, transizioni di stato BOZZA -> IN_REVISIONE -> APPROVATO ->
    PUBBLICATO. Nessuna route di archiviazione/sospensione/bozza-derivata esiste
    ancora (Nota Implementazione 2026-09-16 in tasks.md).
servers:
  - url: /api/v1/builder
security:
  - KeycloakBearer: []
paths:
  /contesti:
    get:
      operationId: listaContestiBuilder
      summary: Contesti autorizzati alla gestione modelli, anche senza integrazioni connesse
      responses:
        '200':
          description: Contesti con permesso gestore derivato separatamente
          content:
            application/json:
              schema: {type: array, items: {type: string}}
              example: [geban]
        '401': {$ref: '#/components/responses/Unauthorized'}
  /tipi-documento/{codiceTipoDocumento}/struttura-disponibile:
    parameters:
      - {$ref: '#/components/parameters/CodiceTipoDocumento'}
    get:
      operationId: getStrutturaDisponibile
      summary: Struttura discovery del tipo documento gia' associato a un'integrazione
      description: >
        Risolve l'integrazione dal codice_contesto di TipoDocumento (non richiede
        di conoscere l'integrazioneId, a differenza delle route /builder/integrazioni/*
        di integrazioni-api.openapi.yaml). Nessun import DB, sempre letto live.
      x-required-role: GEMODO_MODELLI_GESTORE
      responses:
        '200':
          description: Albero discovery corrente del tipo documento
          content:
            application/json:
              schema: {$ref: '../../010-configurazione-cataloghi-integrazioni/contracts/geban-discovery-endpoint.openapi.yaml#/components/schemas/StrutturaTipoDocumento'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '502': {$ref: '#/components/responses/SourceError'}
        '504': {$ref: '#/components/responses/SourceTimeout'}
  /modelli:
    get:
      operationId: listaModelliBuilder
      summary: Modelli e versioni del contesto autorizzato, comprese bozze
      description: Ordinati per creazione decrescente e id. Non richiede discovery online. Nessuna URL amministrativa esposta.
      parameters:
        - name: codice_contesto
          in: query
          required: true
          schema: {type: string, minLength: 1, maxLength: 64}
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, default: 0}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
        # Filtri lato server (007 FR-028). Stanno qui e non nel client perche'
        # la lista e' paginata: filtrare dopo `limit` restituirebbe la pagina in
        # mano invece dell'insieme. Un parametro assente significa "non filtrare".
        - name: codice_tipo_documento
          in: query
          schema: {type: string, minLength: 1, maxLength: 128}
        - name: integrazione_id
          in: query
          schema: {type: string, format: uuid}
        - name: codice_tipologia
          in: query
          schema: {type: string, minLength: 1, maxLength: 128}
        - name: codice_categoria
          in: query
          schema: {type: string, minLength: 1, maxLength: 128}
        - name: lingua
          in: query
          schema: {type: string, enum: [IT, EN]}
        - name: livello_professionale
          in: query
          description: Il token TUTTI seleziona i modelli senza livello, dove la colonna e' nulla.
          schema: {type: string, minLength: 1, maxLength: 64}
        - name: variante
          in: query
          schema: {type: string, minLength: 1, maxLength: 64}
        - name: stato_versione
          in: query
          description: Seleziona i modelli che hanno almeno una versione in questo stato.
          schema:
            type: string
            enum: [BOZZA, IN_REVISIONE, APPROVATO, PUBBLICATO, ARCHIVIATO, SOSPESO]
        - name: ricerca
          in: query
          description: >
            Sottostringa su nome e codice, applicata prima della paginazione.
            Usa ILIKE e non un indice: adeguata a modelli nell'ordine delle
            centinaia, da rivedere se crescessero di ordini di grandezza.
          schema: {type: string, minLength: 1, maxLength: 200}
      responses:
        '200':
          description: Pagina modelli filtrata, vuota se nessun modello corrisponde
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/ModelloGestione'}}
              example: []
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '400': {$ref: '#/components/responses/InvalidInput'}
    post:
      operationId: creaModello
      summary: Crea un modello in BOZZA a partire da un percorso o categoria/tipologia risolti in discovery
      description: >
        Autorizzazione per il singolo contesto del tipo documento target
        (verify_scrittura_su_contesto), mai sull'unione dei contesti del token.
      x-required-role: GEMODO_MODELLI_GESTORE (nel contesto del tipo documento target)
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreaModelloRequest'}
            example:
              codice_tipo_documento: BANDO_CONCORSO
              integrazione_id: 00000000-0000-0000-0000-000000000001
              percorso_categorizzazione: [TD, RICERCATORE]
              livello_professionale: VI
              lingua: IT
      responses:
        '201':
          description: Modello creato in stato BOZZA (senza ancora una versione)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Modello'}
        '400': {$ref: '#/components/responses/InvalidInput'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /tipi-documento/{codiceTipoDocumento}/policy-dimensioni:
    parameters:
      - {$ref: '#/components/parameters/CodiceTipoDocumento'}
    get:
      operationId: getPolicyDimensioni
      summary: Policy registrate e dimensioni ancora prive di policy
      description: >
        DEC-002-POLICY. Dichiara, per nome di dimensione della categorizzazione,
        se ammette un valore generico (un solo modello copre tutti i valori) o se
        ogni valore identifica un modello distinto, come gia' accade per la lingua.
        Esplicitamente scollegata da DefinizioneStruttura di `010`, che e' solo
        l'esempio presentazionale per il team dell'integrazione.
        `dimensioni_non_configurate` alimenta la segnalazione della schermata 4a:
        NodoDiscovery ha extra="allow", quindi senza questo elenco una dimensione
        nuova sparirebbe in silenzio.
      x-required-role: GEMODO_MODELLI_GESTORE
      responses:
        '200':
          description: Policy del tipo documento
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PolicyDimensioni'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      operationId: setPolicyDimensione
      summary: Registra o aggiorna la policy di una dimensione
      description: >
        Una riga sola per (tipo documento, nome dimensione): la policy vale
        ovunque quel nome ricompaia nell'albero, mai per singola foglia.
        Una dimensione la cui persistenza non ammette un valore assente viene
        rifiutata con GENERICO_NON_SUPPORTATO invece di essere accettata e poi
        salvata con un valore arbitrario.
      x-required-role: GEMODO_MODELLI_GESTORE
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PolicyDimensioneRequest'}
      responses:
        '200':
          description: Policy registrata o aggiornata
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PolicyDimensione'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409':
          description: La dimensione non puo' ammettere un valore generico
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Errore'}
              example: {codice: GENERICO_NON_SUPPORTATO, messaggio: "La dimensione 'lingua' non puo' ammettere un valore generico"}
  /modelli/filtri:
    get:
      operationId: vociFiltriModelli
      summary: Valori selezionabili nei filtri dell'elenco modelli
      description: >
        Ricavati dai modelli esistenti nel contesto, non dall'albero discovery:
        l'elenco modelli non richiede discovery online e non deve iniziare a
        dipenderne per riempire delle tendine. Ogni voce restituita produce
        almeno un risultato. Il livello generico compare come TUTTI.
      parameters:
        - name: codice_contesto
          in: query
          required: true
          schema: {type: string, minLength: 1, maxLength: 64}
      responses:
        '200':
          description: Voci selezionabili, liste vuote se il contesto non ha modelli
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VociFiltriModelli'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '400': {$ref: '#/components/responses/InvalidInput'}
  /modelli/{modelloId}:
    parameters:
      - {$ref: '#/components/parameters/ModelloId'}
    get:
      operationId: getModelloDettaglio
      summary: Dettaglio modello con il contratto dati, leggibile anche su una BOZZA
      description: >
        Alimenta l'anteprima del modello nel builder (schermata 2b in versione
        ridotta, 007). A differenza di /catalogo/modelli/{modelloVersioneId}/campi-richiesti,
        che risponde solo su versioni pubblicate, qui i campi sono leggibili anche
        su una versione in BOZZA. Stesso isolamento per contesto della lista.
      x-required-role: GEMODO_MODELLI_GESTORE
      responses:
        '200':
          description: Modello con versioni e campi del contratto
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ModelloDettaglio'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      operationId: eliminaModello
      summary: Elimina logicamente un modello conservando versioni, audit e PDF
      parameters:
        - name: modelloId
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '204': {description: 'Modello eliminato, versioni pubblicate archiviate'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /modelli/{modelloId}/edizioni-derivate:
    parameters:
      - {$ref: '#/components/parameters/ModelloId'}
    post:
      operationId: creaEdizioneDerivata
      summary: Crea un modello collegato in un'altra lingua clonando l'ultima versione
      description: >
        Mantiene tipo, percorso, variante e livello del padre. Crea atomicamente
        un modello indipendente e la sua versione 1 BOZZA copiando formato,
        struttura e campi richiesti dall'ultima versione del padre. Non pubblica
        il derivato e non modifica il padre.
      x-required-role: GEMODO_MODELLI_GESTORE (nel contesto del modello padre)
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreaEdizioneDerivataRequest'}
            example: {lingua: EN}
      responses:
        '201':
          description: Modello derivato e versione 1 BOZZA creati
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ModelloGestione'}
        '400': {$ref: '#/components/responses/InvalidInput'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /modelli/{modelloId}/versioni:
    parameters:
      - {$ref: '#/components/parameters/ModelloId'}
    post:
      operationId: creaVersione
      summary: Crea la versione 1 in BOZZA, campi scelti dalla foglia discovery del modello
      description: >
        I campi devono appartenere al ramo/foglia gia' selezionato dal modello in
        discovery corrente (rifetch aggiornato); nessun campo non ammesso o
        duplicato.
      x-required-role: GEMODO_MODELLI_GESTORE (nel contesto del tipo documento del modello)
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreaVersioneRequest'}
            example:
              campi:
                - {codice: codice_bando, lingua: IT}
                - {codice: titolo_it, lingua: IT}
      responses:
        '201':
          description: Versione creata in stato BOZZA
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Versione'}
        '400': {$ref: '#/components/responses/InvalidInput'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /modelli/{modelloId}/versioni/{versioneId}/invia-revisione:
    parameters:
      - {$ref: '#/components/parameters/ModelloId'}
      - {$ref: '#/components/parameters/VersioneId'}
    post:
      operationId: inviaRevisione
      summary: Transizione BOZZA -> IN_REVISIONE
      x-required-role: GEMODO_MODELLI_GESTORE (nel contesto del tipo documento del modello)
      responses:
        '200':
          description: Versione in stato IN_REVISIONE
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Versione'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /modelli/{modelloId}/versioni/{versioneId}/approva:
    parameters:
      - {$ref: '#/components/parameters/ModelloId'}
      - {$ref: '#/components/parameters/VersioneId'}
    post:
      operationId: approvaVersione
      summary: Transizione IN_REVISIONE -> APPROVATO
      x-required-role: GEMODO_MODELLI_GESTORE (nel contesto del tipo documento del modello)
      responses:
        '200':
          description: Versione in stato APPROVATO
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Versione'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /modelli/{modelloId}/versioni/{versioneId}/pubblica:
    parameters:
      - {$ref: '#/components/parameters/ModelloId'}
      - {$ref: '#/components/parameters/VersioneId'}
    post:
      operationId: pubblicaVersione
      summary: Transizione APPROVATO -> PUBBLICATO
      description: >
        Archivia automaticamente la precedente versione corrente della stessa
        variante, se presente (non restituita in risposta - rileggere il catalogo
        per confermarlo).
      x-required-role: GEMODO_MODELLI_GESTORE (nel contesto del tipo documento del modello)
      responses:
        '200':
          description: Versione in stato PUBBLICATO, ora esposta dal catalogo (001)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Versione'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
components:
  securitySchemes:
    KeycloakBearer: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    CodiceTipoDocumento:
      name: codiceTipoDocumento
      in: path
      required: true
      schema: {type: string, minLength: 1}
...
  schemas:
...
    ModelloGestione:
      type: object
      additionalProperties: false
      required: [id, public_id, codice, nome, codice_tipo_documento, codice_categoria, codice_tipologia, percorso_categorizzazione, variante, dimensioni, lingua, livello_professionale, codice_contesto, integrazione_id, created_at, versioni]
      properties:
        id: {type: string, format: uuid}
        public_id: {type: integer, nullable: true}
        codice: {type: string}
        nome: {type: string}
        codice_tipo_documento: {type: string}
        codice_categoria: {type: string}
        codice_tipologia: {type: string, nullable: true}
        percorso_categorizzazione: {type: array, items: {type: string}}
        variante: {type: string}
        dimensioni:
          type: object
          additionalProperties: {type: string, minLength: 1}
        lingua: {type: string, enum: [IT, EN], nullable: true}
        livello_professionale: {type: string, nullable: true}
        derivato_da_modello_id: {type: string, format: uuid, nullable: true}
        codice_contesto: {type: string}
        integrazione_id: {type: string, format: uuid, nullable: true}
        created_at: {type: string, format: date-time}
        versioni: {type: array, items: {$ref: '#/components/schemas/Versione'}}
...
    Modello:
      type: object
      additionalProperties: false
      required: [id, public_id, codice, nome, codice_tipo_documento, codice_categoria, codice_tipologia, percorso_categorizzazione, variante, lingua, livello_professionale]
      properties:
        id: {type: string, format: uuid}
        public_id: {type: integer, nullable: true}
        codice: {type: string}
        nome: {type: string}
        codice_tipo_documento: {type: string}
        codice_categoria: {type: string}
        codice_tipologia: {type: string, nullable: true}
        percorso_categorizzazione: {type: array, items: {type: string}}
        variante: {type: string}
        lingua: {type: string, enum: [IT, EN]}
        livello_professionale: {type: string, nullable: true}
        derivato_da_modello_id: {type: string, format: uuid, nullable: true}
...
  responses:
    Unauthorized:
      description: JWT assente o non valido
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Errore'}
          example: {codice: ACCESSO_NON_AUTENTICATO, messaggio: Autenticazione richiesta}
    Forbidden:
      description: Ruolo/contesto non autorizzato (mai sull'unione dei contesti del token)
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Errore'}
          example: {codice: ACCESSO_NON_AUTORIZZATO, messaggio: Operazione non autorizzata}
    NotFound:
      description: Modello/versione inesistente, tipo documento non trovato, o ramo non piu' disponibile in discovery
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Errore'}
          examples:
            modello: {value: {codice: MODELLO_NON_TROVATO, messaggio: Modello non trovato}}
            versione: {value: {codice: MODELLO_VERSIONE_NON_TROVATO, messaggio: Versione non trovata}}
            ramo: {value: {codice: CONTESTO_NON_VALIDO, messaggio: Il ramo del modello non e' piu' disponibile}}
    Conflict:
      description: Integrazione non connessa, tipo documento gia' di un'altra integrazione o transizione di stato non valida
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Errore'}
          examples:
            nonConnessa: {value: {codice: INTEGRAZIONE_NON_CONNESSA, messaggio: Integrazione non connessa}}
            transizione: {value: {codice: TRANSIZIONE_STATO_NON_VALIDA, messaggio: Transizione di stato non valida}}
            altraIntegrazione: {value: {codice: TIPO_DOCUMENTO_ALTRA_INTEGRAZIONE, messaggio: Tipo documento gia' associato a un'altra integrazione dello stesso contesto}}
    InvalidInput:
      description: Selezione percorso/categoria incoerente o campi non ammessi/duplicati
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Errore'}
          examples:
            percorso: {value: {codice: CONTESTO_NON_VALIDO, messaggio: Codici incoerenti con il percorso scelto}}
            campi: {value: {codice: CAMPO_NON_AMMESSO, messaggio: Campi duplicati nella versione}}
    SourceError:
      description: Discovery non raggiungibile o non conforme
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Errore'}
          examples:
            trasporto: {value: {codice: DISCOVERY_NON_DISPONIBILE, messaggio: Discovery non disponibile}}
            struttura: {value: {codice: DISCOVERY_NON_CONFORME, messaggio: Discovery non conforme}}
    SourceTimeout:
      description: Budget discovery scaduto
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Errore'}
          example: {codice: DISCOVERY_TIMEOUT, messaggio: Tempo disponibile per discovery scaduto}